escape overwrite at init

+++ init.sh
@@ -8,5 +8,11 @@ if [ ! -d ./$CONTEST ]; then
   mkdir $CONTEST
 fi
 
+if [ ! -e ./$CONTEST/$FNAME ]; then
+  cp template/format.js $CONTEST/$FNAME
+else
+  ESC=$(printf '\033')
+  echo "${ESC}[31mAlready $CONTEST/$FNAME exist${ESC}[m"
+fi
+
 touch $CONTEST/test.txt
-cp template/format.js $CONTEST/$FNAME
